backend/requirements.txt
bcrypt
boto3
cryptography
dataclasses
fastapi
mysql-connector-python
peewee
pycryptodome
pyjwt
python-dotenv
python-jose
python-multipart
six
uuid
uvicorn